Made it so that both testaments can be under one fragment
