+++ app/src/main/java/com/theunquenchedservant/granthornersbiblereadingsystem/ui/settings/BibleStatsFragment.kt
@@ -1,6 +1,7 @@
 package com.theunquenchedservant.granthornersbiblereadingsystem.ui.settings
 
 import android.os.Bundle
+import androidx.core.os.bundleOf
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import com.theunquenchedservant.granthornersbiblereadingsystem.MainActivity
@@ -14,11 +15,13 @@ class BibleStatsFragment : PreferenceFragmentCompat() {
         val bibleRead: Preference? = findPreference("bibleRead")
         val mainActivity = activity as MainActivity
         oldTestament!!.onPreferenceClickListener = Preference.OnPreferenceClickListener {
-            mainActivity.navController.navigate(R.id.navigation_bible_stats_old)
+            val bundle = bundleOf("testament" to "old")
+            mainActivity.navController.navigate(R.id.navigation_bible_testament_stats, bundle)
             false
         }
         newTestament!!.onPreferenceClickListener = Preference.OnPreferenceClickListener {
-            mainActivity.navController.navigate(R.id.navigation_bible_stats_new)
+            val bundle = bundleOf("testament" to "new")
+            mainActivity.navController.navigate(R.id.navigation_bible_testament_stats, bundle)
             false
         }
     }
