+++ app/src/main/res/xml/bible_statistics_main.xml
@@ -6,17 +6,21 @@
         android:focusableInTouchMode="true"
         app:key="oldTestament"
         app:summary="0%"
-        app:title="@string/title_old_testament_read" />
+        app:title="@string/title_old_testament_read"
+        app:allowDividerBelow="true"/>
     <Preference
         android:clickable="true"
         android:focusableInTouchMode="true"
         app:key="newTestament"
         app:summary="0%"
-        app:title="@string/title_new_testament_read" />
+        app:title="@string/title_new_testament_read"
+        app:allowDividerAbove="true"
+        app:allowDividerBelow="true"/>
     <Preference
         android:clickable="false"
         android:focusableInTouchMode="false"
         android:summary="0"
         app:key="bibleRead"
-        app:title="@string/title_bible_read"/>
+        app:title="@string/title_bible_read"
+        app:allowDividerAbove="true" />
 </PreferenceScreen>
