--- app/src/main/res/xml/new_testament_stats.xml
@@ -1,166 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="matthew"
-        app:summary="0% | Times Read: 0"
-        app:title="Matthew" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="mark"
-        app:summary="0% | Times Read: 0"
-        app:title="Mark" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="luke"
-        app:summary="0% | Times Read: 0"
-        app:title="Luke" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="john"
-        app:summary="0% | Times Read: 0"
-        app:title="John" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="acts"
-        app:summary="0% | Times Read: 0"
-        app:title="Acts" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="romans"
-        app:summary="0% | Times Read: 0"
-        app:title="Romans" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="corinthians1"
-        app:summary="0% | Times Read: 0"
-        app:title="1 Corinthians" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="corinthians2"
-        app:summary="0% | Times Read: 0"
-        app:title="2 Corinthians" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="galatians"
-        app:summary="0% | Times Read: 0"
-        app:title="Galatians" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="ephesians"
-        app:summary="0% | Times Read: 0"
-        app:title="Ephesians" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="philippians"
-        app:summary="0% | Times Read: 0"
-        app:title="Philippians" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="colossians"
-        app:summary="0% | Times Read: 0"
-        app:title="Colossians" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="thessalonians1"
-        app:summary="0% | Times Read: 0"
-        app:title="1 Thessalonians" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="thessalonians2"
-        app:summary="0% | Times Read: 0"
-        app:title="2 Thessalonians" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="timothy1"
-        app:summary="0% | Times Read: 0"
-        app:title="1 Timothy" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="timothy2"
-        app:summary="0% | Times Read: 0"
-        app:title="2 Timothy" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="titus"
-        app:summary="0% | Times Read: 0"
-        app:title="Titus" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="philemon"
-        app:summary="0% | Times Read: 0"
-        app:title="Philemon" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="hebrews"
-        app:summary="0% | Times Read: 0"
-        app:title="Hebrews" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="james"
-        app:summary="0% | Times Read: 0"
-        app:title="James" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="peter1"
-        app:summary="0% | Times Read: 0"
-        app:title="1 Peter" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="peter2"
-        app:summary="0% | Times Read: 0"
-        app:title="2 Peter" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="john1"
-        app:summary="0% | Times Read: 0"
-        app:title="1 John" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="john2"
-        app:summary="0% | Times Read: 0"
-        app:title="2 John" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="john3"
-        app:summary="0% | Times Read: 0"
-        app:title="3 John" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="jude"
-        app:summary="0% | Times Read: 0"
-        app:title="Jude" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="revelation"
-        app:summary="0% | Times Read: 0"
-        app:title="Revelation" />
-</PreferenceScreen>
