--- app/src/main/res/xml/old_testament_stats.xml
@@ -1,238 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="genesis"
-        app:summary="0% | Times Read: 0"
-        app:title="Genesis" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="exodus"
-        app:summary="0% | Times Read: 0"
-        app:title="Exodus" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="leviticus"
-        app:summary="0% | Times Read: 0"
-        app:title="Leviticus" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="numbers"
-        app:summary="0% | Times Read: 0"
-        app:title="Numbers" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="deuteronomy"
-        app:summary="0% | Times Read: 0"
-        app:title="Deuteronomy" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="joshua"
-        app:summary="0% | Times Read: 0"
-        app:title="Joshua" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="judges"
-        app:summary="0% | Times Read: 0"
-        app:title="Judges" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="ruth"
-        app:summary="0% | Times Read: 0"
-        app:title="Ruth" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="samuel1"
-        app:summary="0% | Times Read: 0"
-        app:title="1 Samuel" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="samuel2"
-        app:summary="0% | Times Read: 0"
-        app:title="2 Samuel" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="kings1"
-        app:summary="0% | Times Read: 0"
-        app:title="1 Kings" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="kings2"
-        app:summary="0% | Times Read: 0"
-        app:title="2 Kings" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="chronicles1"
-        app:summary="0% | Times Read: 0"
-        app:title="1 Chronicles" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="chronicles2"
-        app:summary="0% | Times Read: 0"
-        app:title="2 Chronicles" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="ezra"
-        app:summary="0% | Times Read: 0"
-        app:title="Ezra" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="nehemiah"
-        app:summary="0% | Times Read: 0"
-        app:title="Nehemiah" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="esther"
-        app:summary="0% | Times Read: 0"
-        app:title="Esther" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="job"
-        app:summary="0% | Times Read: 0"
-        app:title="Job" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="psalm"
-        app:summary="0% | Times Read: 0"
-        app:title="Psalm" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="proverbs"
-        app:summary="0% | Times Read: 0"
-        app:title="Proverbs" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="ecclesiastes"
-        app:summary="0% | Times Read: 0"
-        app:title="Ecclesiastes" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="song"
-        app:summary="0% | Times Read: 0"
-        app:title="Song of Solomon" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="isaiah"
-        app:summary="0% | Times Read: 0"
-        app:title="Isaiah" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="jeremiah"
-        app:summary="0% | Times Read: 0"
-        app:title="Jeremiah" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="lamentations"
-        app:summary="0% | Times Read: 0"
-        app:title="Lamentations" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="ezekiel"
-        app:summary="0% | Times Read: 0"
-        app:title="Ezekiel" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="daniel"
-        app:summary="0% | Times Read: 0"
-        app:title="Daniel" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="hosea"
-        app:summary="0% | Times Read: 0"
-        app:title="Hosea" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="joel"
-        app:summary="0% | Times Read: 0"
-        app:title="Joel" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="amos"
-        app:summary="0% | Times Read: 0"
-        app:title="Amos" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="obadiah"
-        app:summary="0% | Times Read: 0"
-        app:title="Obadiah" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="jonah"
-        app:summary="0% | Times Read: 0"
-        app:title="Jonah" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="micah"
-        app:summary="0% | Times Read: 0"
-        app:title="Micah" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="nahum"
-        app:summary="0% | Times Read: 0"
-        app:title="Nahum" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="habakkuk"
-        app:summary="0% | Times Read: 0"
-        app:title="Habakkuk" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="zephaniah"
-        app:summary="0% | Times Read: 0"
-        app:title="Zephaniah" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="haggai"
-        app:summary="0% | Times Read: 0"
-        app:title="Haggai" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="zechariah"
-        app:summary="0% | Times Read: 0"
-        app:title="Zechariah" />
-    <Preference
-        android:clickable="true"
-        android:focusableInTouchMode="true"
-        app:key="malachi"
-        app:summary="0% | Times Read: 0"
-        app:title="Malachi" />
-</PreferenceScreen>
